https://extract.atlassian.net/browse/ISSUE-13503 Modified catch when opening database to rethrow exception instead of eating it
https://extract.atlassian.net/browse/ISSUE-1163
Modified code to make sure the placeholder attribute gets placed at the end of the attributes for group when column set to “.” and it isn’t the first column in a table. Also fixed problem of a hidden column mapped to . causing an exception to be thrown.

Reviewed by Nat(1/0)

William

[git-vault-id] Extract$/Engineering@36582/118771

+++ RC.Net/UtilityApplications/SQLCompactUtils/SQLCDBEditor/Core/Code/SQLCDBEditorForm.cs
@@ -1990,13 +1990,18 @@ namespace Extract.SQLCDBEditor
                     // Open the connection.
                     _connection.Open();
                 }
-                catch
+                catch (Exception ex)
                 {
                     // If we failed to open a new database connection, delete the working database
                     // file and set _databaseWorkingCopyFileName to null to ensure this instance
                     // doesn't end up trying to delete a working copy created by another instance.
-                    FileSystemMethods.DeleteFile(_databaseWorkingCopyFileName);
+                    // Verify the file exists before attempting to delete it
+                    if (File.Exists(_databaseWorkingCopyFileName))
+                    {
+                        FileSystemMethods.DeleteFile(_databaseWorkingCopyFileName);
+                    }
                     _databaseWorkingCopyFileName = null;
+                    throw ex.AsExtract("ELI39234");
                 }
 
                 CheckSchemaVersionAndPromptForUpdate();
